requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
colorama>=0.4.6
python-nmap>=0.7.1
tqdm>=4.66.0
lxml>=4.9.3